moved more functions to util
